+++ React/src/app/pages/game/components/Board/Board.module.scss
@@ -27,7 +27,7 @@
   border: 1px dashed #d9d9d9;
   height: 96px;
   display: flex;
-  flex: 0 0 96px;
+  flex-basis: 96px;
   justify-content: center;
   align-items: center;
   cursor: pointer;
